finally connected to db
