--- EVCP.RouteFInding/TestClass.cs
@@ -1,11 +1,0 @@
-﻿namespace EVCP.RouteFInding;
-
-public class TestClass
-{
-
-    
-    public static void Main(string[] args){
-        Console.WriteLine("hello im alive and well");
-
-    }
-}
