+++ EVCP.RouteFinding/appsettings.json
@@ -1,0 +1,12 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "ConnectionStrings": {
+    "EVDataWarehouse": "Server=127.0.0.1;Port=5432;Database=EVDataWarehouse;User Id=sobibence;Password=sobibence;"
+  },
+  "AllowedHosts": "*"
+}
